Extract shared code to BaseHighlighter
