+++ app/assets/javascripts/hera.js
@@ -21,6 +21,7 @@
 //= require hera/modules/console_updater
 //= require hera/modules/differ
 //= require hera/modules/editor_toolbar
+//= require hera/modules/base_highlighter
 //= require hera/modules/grammar_highlighter
 //= require hera/modules/quote_selector
 //= require hera/modules/inline_thread_utils
