+++ engines/dradis-echo/app/assets/config/dradis/plugins/echo/manifests/hera.js
@@ -1,5 +1,4 @@
 //= link controllers/dradis/plugins/echo/agent_controller.js
-//= link controllers/dradis/plugins/echo/configurations_controller.js
 //= link controllers/dradis/plugins/echo/grammar_controller.js
 //= link controllers/dradis/plugins/echo/prompt_controller.js
 //= link controllers/dradis/plugins/echo/prompt_selector_controller.js
